Fix folder capitalization, moved seperate code component folders into 1 main folder
