+++ src/App.js
@@ -1,19 +1,19 @@
 import './App.css';
-import Header from './components/GeneralComponents/Header';
+import Header from './components/general/Header';
 import AboutUs from './components/AboutUs';
 import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
-import AssignmentsPage from './components/AssignmentComponents/AssignmentsPage';
-import UploadProjectPage from './components/CodeUploadComponents/UploadMethodList';
+import AssignmentsPage from './components/assignment/AssignmentsPage';
+import UploadProjectPage from './components/code/codeUpload/UploadMethodList';
 import HomePage from './components/HomePage';
-import Footer from './components/GeneralComponents/Footer';
-import UploadMethodList from './components/CodeUploadComponents/UploadMethodList';
-import GitUpload from './components/CodeUploadComponents/GitUploadPage';
-import FolderUpload from './components/CodeUploadComponents/FolderUploadPage';
-import AssignmentInProgress from './components/AssignmentComponents/AssignmentsInProgress';
-import Login from './components/LoginComponents/Login';
-import Course from './components/CourseComponents/Course';
-import CodeFeedbackPage from './components/CodeFeedbackComponents/CodeFeedbackPage';
-import AssignmentWorkspace from './components/AssignmentComponents/AssignmentWorkspace';
+import Footer from './components/general/Footer';
+import UploadMethodList from './components/code/codeUpload/UploadMethodList';
+import GitUpload from './components/code/codeUpload/GitUploadPage';
+import FolderUpload from './components/code/codeUpload/FolderUploadPage';
+import AssignmentInProgress from './components/assignment/AssignmentsInProgress';
+import Login from './components/login/Login';
+import Course from './components/course/Course';
+import CodeFeedbackPage from './components/code/codeFeedback/CodeFeedbackPage';
+import AssignmentWorkspace from './components/assignment/AssignmentWorkspace';
 import {AuthProvider, useAuth} from "./components/context/AuthContext";
 import AuthorizedRoute from "./components/security/AuthorizedRoute";
 import PrivateRoute from "./components/security/PrivateRoute";
