+++ src/components/assignment/AssignmentWorkspace.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import CodeField from '../CodeFeedbackComponents/Codefield';
+import CodeField from '../code/codeFeedback/Codefield';
 import CodeService from '../../services/CodeService';
 import { useParams } from 'react-router-dom';
 import '../../style/AssignmentWorkspace.css';
